requirements.txt
numpy==1.20.3
scikit-learn==0.24.2
tensorflow==2.6.0
Pillow==8.4.0
Flask==2.1.1
